@@ -11,17 +11,17 @@
   "postcss:watch": "postcss src/style.css --use postcss-import --dir public --watch",
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.10",
+    "@angular/animations": "^18.2.13",
     "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.3.10",
-    "@angular/compiler": "^17.3.10",
-    "@angular/core": "^17.3.10",
+    "@angular/common": "^18.2.13",
+    "@angular/compiler": "^18.2.13",
+    "@angular/core": "^18.2.13",
     "@angular/fire": "^17.1.0",
-    "@angular/forms": "^17.3.10",
+    "@angular/forms": "^18.2.13",
     "@angular/material": "^17.0.0",
-    "@angular/platform-browser": "^17.3.10",
-    "@angular/platform-browser-dynamic": "^17.3.10",
-    "@angular/router": "^17.3.10",
+    "@angular/platform-browser": "^18.2.13",
+    "@angular/platform-browser-dynamic": "^18.2.13",
+    "@angular/router": "^18.2.13",
     "@manhnd/esc-pos-encoder": "^1.0.25",
     "@ng-bootstrap/ng-bootstrap": "^16.0.0",
     "@ng-select/ng-select": "^12.0.0",
@@ -62,10 +62,10 @@
   },
   "devDependencies": {
     "@angular-builders/custom-webpack": "^17.0.0",
-    "@angular-devkit/build-angular": "^17.0.0",
-    "@angular/cli": "^18.2.11",
-    "@angular/compiler-cli": "^17.3.10",
-    "@angular/localize": "^17.3.10",
+    "@angular-devkit/build-angular": "^18.2.15",
+    "@angular/cli": "^18.2.15",
+    "@angular/compiler-cli": "^18.2.13",
+    "@angular/localize": "^18.2.13",
     "@types/crypto-js": "^4.2.2",
     "@types/jasmine": "~4.3.0",
     "@types/lodash": "^4.17.13",
